audit-app.py
import streamlit as st
import os
import xml.etree.ElementTree as ET
import pandas as pd
from streamlit_tree_select import tree_select

from audit_reformat import handle_audit_reformat

#This script is intended to be an end-to-end audit powered by LLMs
#It will start with uploading the audit output from Qualtrics,
#then will present the user an interface to allow them to select what part of the model they want audited,
#then peform the audit using an LLM and return the completed audit
# along with accuracy, summary of findings, and suggestions for improvement per category

# Configure Streamlit page
st.set_page_config(page_title="Enhanced Audit")

def main():
    # st.cache_data.clear()
    st.title("Enhanced Audit Script")
    st.write("Run an audit using an LLM to determine accuracy.")

    uploaded_file = st.file_uploader(
        "Upload raw audit file (.xlsx)",
        type=["xlsx"],
    )

    if st.button("Process", type="primary"):
        if uploaded_file is None:
            st.error("Please upload an audit file before processing.")
            return
        with st.spinner("Reformatting audit..."):
            output = handle_audit_reformat(uploaded_file)
            st.success("Reformat complete.")
            st.download_button(
                label="Download reformatted audit",
                data=output.getvalue(),
                file_name="reformatted_audit.xlsx",
                mime="application/vnd.openxmlformats-officedocument.spreadsheetml.sheet",
            )

if __name__ == "__main__":
    main()

#allow upload of raw audit file

#allow upload of model XML file to provide descriptions for LLM

#allow user to select which branches they want the audit to apply to

#allow user to specify audit prompt

#reformat the audit for analysis e.g. to unmerge cells

#pull together extra category info to send do the LLM, e.g. full category tree, category description

#send the sentences to the LLM to judge and return judgments

#Also have the LLM summarize the main issues by category and suggest rules improvements

#assemble audited sentences back into a spreadsheet

#set up the spreadsheet to report category accuracy

#output spreadsheet
