requirements.txt
anthropic==0.75.0
openpyxl==3.1.5
pandas==2.3.3
python-dotenv==1.2.1
PyYAML==6.0.3
streamlit==1.52.2
xlsxwriter==3.2.9
